modafied postcard querry
