+++ server/config/connection.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
 
-mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost:27017/GamersGuild');
+mongoose.connect(process.env.MONGODB_URI || 'mongodb://127.0.0.1:27017/GamersGuild');
 
 module.exports = mongoose.connection;
